now it compiles

+++ srcs/free_deallocate_small.c
@@ -6,7 +6,7 @@
 /*   By: cledant <cledant@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/01/19 17:51:46 by cledant           #+#    #+#             */
-/*   Updated: 2017/01/19 17:53:16 by cledant          ###   ########.fr       */
+/*   Updated: 2017/01/19 18:32:26 by cledant          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@ int		free_deallocate_small(t_small *header, const short int user_id,
 
 	i = alloc_id;
 	j = 0;
-	if (header->used_alloc < nb_alloc)
+	if ((size_t)header->used_alloc < nb_alloc)
 		return (FREE_NOP);
 	header->used_alloc -= nb_alloc;
 	(header->index)[user_id - 1] = NOT_USED;
